+++ src/LiveChartsCore/LiveCharts.cs
@@ -103,12 +103,12 @@ public static class LiveCharts
     /// <summary>
     /// Gets a constant that indicates that a rotation angle follows the tangent line, this property is only useful in polar series.
     /// </summary>
-    public static int TangentAngle => 1 << 25;
+    public static double TangentAngle => 1 << 25;
 
     /// <summary>
     /// Gets a constant that indicates that a rotation angle follows the cotangent line, this property is only useful in polar series.
     /// </summary>
-    public static int CotangentAngle => 1 << 26;
+    public static double CotangentAngle => 1 << 26;
 
     /// <summary>
     /// The disable animations
